Christopher Wandel - Interview Submission
I left comments throughout the code for some explanation. I wrote a console app that writes my answers to the console and a text file. It is set up to read all the text files in the 'Data' folder, and it sends the results to 'Output/answers.txt'. 

Please let me know if any questions or clarification is needed. Thank you!